replace comments into murr_card app
