--- comments/admin.py
@@ -1,7 +1,0 @@
-from django.contrib import admin
-
-from comments.models import Comment
-
-
-admin.site.register(Comment)
-
