--- comments/serializers.py
@@ -1,13 +1,0 @@
-from rest_framework import serializers
-
-from comments.models import Comment
-
-
-class CommentSerializer(serializers.ModelSerializer):
-    owner_name = serializers.ReadOnlyField(source='owner.username')
-
-    class Meta:
-        model = Comment
-        fields = ('owner', 'murr_card', 'text', 'owner_name', 'depth', 'parent')
-
-
