+++ murr_back/urls.py
@@ -13,6 +13,5 @@ urlpatterns = [
     # local
     path('api/murren/', include('murren.urls')),
     path('api/murr_card/', include('murr_card.urls')),
-    path('api/comments/', include('comments.urls')),
 
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
