+++ murr_card/admin.py
@@ -1,4 +1,5 @@
 from django.contrib import admin
-from .models import MurrCard
+from .models import MurrCard, Comment
 
 admin.site.register(MurrCard)
+admin.site.register(Comment)
