+++ murr_card/serializers.py
@@ -1,7 +1,14 @@
 from rest_framework import serializers
 
-from comments.serializers import CommentSerializer
-from .models import MurrCard, EditorImageForMurrCard
+from .models import MurrCard, EditorImageForMurrCard, Comment
+
+
+class CommentSerializer(serializers.ModelSerializer):
+    owner_name = serializers.ReadOnlyField(source='owner.username')
+
+    class Meta:
+        model = Comment
+        fields = ('owner', 'murr_card', 'text', 'owner_name', 'depth', 'parent')
 
 
 class MurrCardSerializers(serializers.ModelSerializer):
